refactor: polish chain and dependency package
