+++ api/pybot/chains/__init__.py
@@ -1,0 +1,3 @@
+from pybot.chains.summarization import SummarizationChain
+
+__all__ = ["SummarizationChain"]
