+++ api/pybot/dependencies.py
@@ -1,6 +1,8 @@
 from typing import Annotated, Optional
 
 from fastapi import Depends, Header
+from langchain.agents import AgentExecutor
+from langchain.chains.base import Chain
 from langchain_community.chat_message_histories import RedisChatMessageHistory
 from langchain_community.llms.huggingface_text_gen_inference import (
     HuggingFaceTextGenInference,
@@ -8,11 +10,14 @@ from langchain_community.llms.huggingface_text_gen_inference import (
 from langchain_core.language_models import BaseLLM
 from langchain_core.memory import BaseMemory
 
+from pybot.agent.base import create_agent
 from pybot.callbacks import TracingLLMCallbackHandler
+from pybot.chains import SummarizationChain
 from pybot.config import settings
 from pybot.history import PybotMessageHistory
 from pybot.memory import PybotMemory
 from pybot.prompts.chatml import AI_SUFFIX
+from pybot.tools import CodeSandbox
 
 
 def UserIdHeader(alias: Optional[str] = None, **kwargs):
@@ -64,4 +69,33 @@ def Llm() -> BaseLLM:
         ],  # not all mistral models have a decent tokenizer config.
         streaming=True,
         callbacks=[TracingLLMCallbackHandler()],
+    )
+
+
+def PybotAgent(
+    llm: Annotated[BaseLLM, Depends(Llm)],
+    memory: Annotated[BaseMemory, Depends(ChatMemory)],
+) -> AgentExecutor:
+    tools = [
+        CodeSandbox(
+            gateway_url=str(settings.jupyter_enterprise_gateway_url),
+        )
+    ]
+    return create_agent(
+        llm=llm,
+        tools=tools,
+        agent_executor_kwargs={
+            "memory": memory,
+            "return_intermediate_steps": True,
+        },
+    )
+
+
+def SmryChain(
+    llm: Annotated[BaseLLM, Depends(Llm)],
+    memory: Annotated[BaseMemory, Depends(ChatMemory)],
+) -> Chain:
+    return SummarizationChain(
+        llm=llm,
+        memory=memory,
     )
