+++ api/pybot/routers/conversations.py
@@ -1,15 +1,14 @@
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException
+from langchain.chains.base import Chain
 from langchain_community.chat_message_histories import RedisChatMessageHistory
-from langchain_core.language_models import BaseLLM
-from langchain_core.memory import BaseMemory
 from langchain_core.messages import HumanMessage
 from loguru import logger
 
 from pybot.config import settings
 from pybot.context import Session, session_id
-from pybot.dependencies import ChatMemory, Llm, MessageHistory, UserIdHeader
+from pybot.dependencies import MessageHistory, SmryChain, UserIdHeader
 from pybot.jupyter import ContextAwareKernelManager, GatewayClient
 from pybot.models import Conversation as ORMConversation
 from pybot.schemas import (
@@ -19,7 +18,6 @@ from pybot.schemas import (
     CreateConversation,
     UpdateConversation,
 )
-from pybot.summarization import summarize as summarize_conv
 
 router = APIRouter(
     prefix="/api/conversations",
@@ -123,15 +121,15 @@ async def delete_conversation(
 @router.post("/{conversation_id}/summarization", status_code=201)
 async def summarize(
     conversation_id: str,
-    llm: Annotated[BaseLLM, Depends(Llm)],
-    memory: Annotated[BaseMemory, Depends(ChatMemory)],
+    smry_chain: Annotated[Chain, Depends(SmryChain)],
     userid: Annotated[str | None, UserIdHeader()] = None,
 ) -> dict[str, str]:
     conv = await ORMConversation.get(conversation_id)
     if conv.owner != userid:
         raise HTTPException(status_code=403, detail="authorization error")
     session_id.set(f"{userid}:{conversation_id}")
-    title = await summarize_conv(llm, memory)
+    res = await smry_chain.ainvoke(input={})
+    title = res[smry_chain.output_key]
     conv.title = title
     await conv.save()
     return {"title": title}
